refactor: переименовать файлы моделей (models.py) в модулях comments и users
