+++ backend/blogapp/core/database/mongodb.py
@@ -3,8 +3,8 @@ from motor.motor_asyncio import AsyncIOMotorClient
 
 from ..config import MONGODB_URL, MONGODB_DATABASE
 from ...modules.articles.models import ArticleDocument
-from ...modules.comments.model import CommentDocument
-from ...modules.users.model import UserDocument
+from ...modules.comments.models import CommentDocument
+from ...modules.users.models import UserDocument
 
 
 # Init beanie ODM
