+++ backend/blogapp/core/security/routes.py
@@ -9,7 +9,7 @@ from starlette import status
 
 from .schema import TokenResponseBody, RegisterRequestBody
 from .utilities import authenticate_user, create_access_token, get_password_hash
-from ...modules.users.model import UserDocument
+from ...modules.users.models import UserDocument
 
 router = APIRouter()
