+++ backend/blogapp/modules/articles/models.py
@@ -5,7 +5,7 @@ from typing import Annotated
 from beanie import Link
 from pydantic import BaseModel, model_validator, Field
 
-from ..users.model import UserDocument
+from ..users.models import UserDocument
 from ...core.database.extended_document import ExtendedDocument
